requirements.txt
fastapi
numpy
pandas
protobuf==3.19.6
python-dotenv
rapidfuzz
SPARQLWrapper
stanza==1.3.0
starlette
uvicorn